+++ lib/assets/constants.dart
@@ -1,5 +1,5 @@
 class PaylinkConstants {
-  /// All payment methods are accepted by Paylink
+  // Valid card brands accepted by Paylink.
   static const List<String> validCardBrands = [
     'mada',
     'visaMastercard',
@@ -10,17 +10,17 @@ class PaylinkConstants {
     'urpay'
   ];
 
-  /// Testing API credentials.
-  static const String testingApiId = 'APP_ID_1123453311';
-  static const String testingSecretKey = '0662abb5-13c7-38ab-cd12-236e58f43766';
+  // API URLs for production and test environments
+  static const String testApiBaseUrl = 'https://restpilot.paylink.sa';
+  static const String productionApiBaseUrl = 'https://restapi.paylink.sa';
 
-  /// API Link
-  static const String testApiLink = 'https://restpilot.paylink.sa';
-  static const String productionApiLink = 'https://restapi.paylink.sa';
-
-  /// Payment Frame Url
+  // Payment Frame URLs for production and test environments
   static const String testingPaymentFrameUrl =
       'https://paymentpilot.paylink.sa/pay/frame';
   static const String productionPaymentFrameUrl =
       'https://payment.paylink.sa/pay/frame';
+
+  // Default credentials for the test environment
+  static const String testingApiId = 'APP_ID_1123453311';
+  static const String testingSecretKey = '0662abb5-13c7-38ab-cd12-236e58f43766';
 }