+++ lib/paylink_payment.dart
@@ -2,10 +2,14 @@ library paylink_payment;
 
 import 'package:flutter/material.dart';
 import 'package:paylink_payment/api/paylink_api.dart';
+import 'package:paylink_payment/assets/helpers.dart';
 import 'package:paylink_payment/widgets/payment_webview.dart';
+import 'package:paylink_payment/assets/constants.dart';
+import 'package:paylink_payment/models/paylink_invoice.dart';
 
 // export PaylinkProduct class
 export 'package:paylink_payment/models/paylink_product.dart';
+export 'package:paylink_payment/models/paylink_invoice.dart';
 
 /// A class for handling payments using PaylinkAPI.
 class PaylinkPayment extends PaylinkAPI {
@@ -17,23 +21,57 @@ class PaylinkPayment extends PaylinkAPI {
   final Color textColor;
   final Color themeColor;
 
-  /// Creates a new [PaylinkPayment] instance.
-  ///
-  /// [isTestMode] - Indicates whether the API is in test mode.
-  /// [apiId] - Production API ID.
-  /// [secretKey] - Production API secret key.
-  /// [webViewTitle] - The title to display in the payment web view.
-  /// [textColor] - The text color for the payment web view.
-  /// [themeColor] - The theme color for the payment web view.
-  PaylinkPayment({
-    super.isTestMode,
+  /// Private constructor
+  PaylinkPayment._internal({
+    required this.context,
+    required this.webViewTitle,
+    required this.textColor,
+    required this.themeColor,
     required super.apiId,
     required super.secretKey,
-    required this.context,
-    this.webViewTitle = 'Payment',
-    this.textColor = Colors.white,
-    this.themeColor = const Color(0xFF6a64ef),
+    required super.apiBaseUrl,
+    required super.paymentFrameUrl,
   });
+
+  /// Factory constructor for test environment
+  factory PaylinkPayment.test({
+    required BuildContext context,
+    String webViewTitle = 'Payment',
+    Color textColor = Colors.white,
+    Color themeColor = const Color(0xFF6a64ef),
+  }) {
+    return PaylinkPayment._internal(
+      context: context,
+      webViewTitle: webViewTitle,
+      textColor: textColor,
+      themeColor: themeColor,
+      apiId: PaylinkConstants.testingApiId,
+      secretKey: PaylinkConstants.testingSecretKey,
+      apiBaseUrl: PaylinkConstants.testApiBaseUrl,
+      paymentFrameUrl: PaylinkConstants.testingPaymentFrameUrl,
+    );
+  }
+
+  /// Factory constructor for production environment
+  factory PaylinkPayment.production({
+    required BuildContext context,
+    required String apiId,
+    required String secretKey,
+    String webViewTitle = 'Payment',
+    Color textColor = Colors.white,
+    Color themeColor = const Color(0xFF6a64ef),
+  }) {
+    return PaylinkPayment._internal(
+      context: context,
+      webViewTitle: webViewTitle,
+      textColor: textColor,
+      themeColor: themeColor,
+      apiId: apiId,
+      secretKey: secretKey,
+      apiBaseUrl: PaylinkConstants.productionApiBaseUrl,
+      paymentFrameUrl: PaylinkConstants.productionPaymentFrameUrl,
+    );
+  }
 
   /// Opens the payment form in a web view.
   ///
@@ -43,29 +81,23 @@ class PaylinkPayment extends PaylinkAPI {
   void openPaymentForm({
     required String transactionNo,
     required final Function(
-      Map<String, dynamic> orderDetails,
+      PaylinkInvoice orderDetails,
     ) onPaymentComplete,
     required final Function(Object error) onError,
   }) async {
     try {
       // Fetch order details from the API
-      var orderDetails = await getInvoice(transactionNo: transactionNo);
+      var orderDetails = await getInvoice(transactionNo);
 
       // Extract gateway order request from order details
-      Map<String, dynamic>? gatewayOrderRequest =
-          orderDetails['gatewayOrderRequest'];
-
-      // Check if gateway order request is available
-      if (gatewayOrderRequest == null) {
-        throw Exception('No gateway order request found in order details');
-      }
+      final gatewayOrderRequest = orderDetails.gatewayOrderRequest;
 
       // Show the payment form in a web view
       _showWebView(
-        orderDetails['transactionNo'],
-        gatewayOrderRequest['clientName'],
-        gatewayOrderRequest['clientMobile'],
-        gatewayOrderRequest['callBackUrl'],
+        orderDetails.transactionNo,
+        gatewayOrderRequest.clientName,
+        gatewayOrderRequest.clientMobile,
+        gatewayOrderRequest.callBackUrl,
         onPaymentComplete,
       );
     } catch (e) {
@@ -85,7 +117,7 @@ class PaylinkPayment extends PaylinkAPI {
     String? clientName,
     String? clientMobile,
     String? callBackUrl,
-    final Function(Map<String, dynamic> orderDetails) onPaymentComplete,
+    final Function(PaylinkInvoice orderDetails) onPaymentComplete,
   ) {
     try {
       if (transactionNo == null) {
@@ -97,10 +129,11 @@ class PaylinkPayment extends PaylinkAPI {
       }
 
       // Construct the payment page URL
-      String paymentPageUrl = getPaymentPageUrl(
+      String paymentPageUrl = PaylinkHelper.getPaymentPageUrl(
+        paymentFrameUrl,
         transactionNo,
-        clientName,
-        clientMobile,
+        clientName ?? '',
+        clientMobile ?? '',
       );
 
       // Navigate to the payment web view
@@ -131,7 +164,7 @@ class PaylinkPayment extends PaylinkAPI {
   /// [onPaymentComplete] - A callback function to handle the payment result.
   void handleCallbackReached(
     Map<String, String> callBackParams,
-    final Function(Map<String, dynamic> orderDetails) onPaymentComplete,
+    final Function(PaylinkInvoice orderDetails) onPaymentComplete,
   ) async {
     try {
       // Extract transaction number from callback parameters
@@ -146,16 +179,16 @@ class PaylinkPayment extends PaylinkAPI {
       }
 
       // Fetch order details using the transaction number
-      var orderDetails = await getInvoice(transactionNo: transactionNo);
+      final orderDetails = await getInvoice(transactionNo);
 
       // Extract order amount and order status from order details
-      String? orderStatus = orderDetails['orderStatus'];
-      Map<String, dynamic>? paymentErrors = paymentErrorsToMap(
-        orderDetails['paymentErrors'],
+      String orderStatus = orderDetails.orderStatus;
+      Map<String, dynamic> paymentErrors = PaylinkHelper.paymentErrorsToMap(
+        orderDetails.paymentErrors,
       );
 
       // Throw an error if the payment is not paid
-      if (orderStatus?.toLowerCase() != 'paid') {
+      if (orderStatus.toLowerCase() != 'paid') {
         String errorMsg = paymentErrors['errorTitle'] ??
             'Failed to proccess the payment, try again!';
         throw Exception(errorMsg);